+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 -   Add new functions `OlmMachine.export_secrets_for_secret_storage` and
     `OlmMachine.import_secrets_from_secret_storage`, along with associated
     types, for reading and writing to Secret Storage. [#67](https://github.com/matrix-org/matrix-rust-sdk-crypto-nodejs/pull/67)
+-   Add a new function `OlmMachine.get_device` to get information about a
+    device. [#67](https://github.com/matrix-org/matrix-rust-sdk-crypto-nodejs/pull/67)
 
 ## v0.4.0 - 2026-01-08